==> 01-data_prep-THD assignment.ipynb <==
import os
os.chdir('/content/drive/MyDrive/Colab Notebooks/Technical Assessement THD')
print(os.getcwd())

import pandas as pd
import numpy as np
import datetime


df=pd.read_csv('store_sales_data_final.csv')


df_base=df.drop(['store_info','holiday_info'],axis=1)
df_base



df1=df[['date','store_nbr','store_info']].reset_index(drop=True)

df1['store_info']=df1['store_info'].map(eval)


df_store = pd.json_normalize(df1['store_info'])

df1=pd.concat([df1,df_store],axis=1)

processed_store=df1.drop(['store_info'],axis=1)

processed_store





df_holiday_not_null=df[~df['holiday_info'].isnull()].reset_index(drop=True)

cols= ['date','store_nbr','holiday_info']

df_holiday_not_null=df_holiday_not_null[cols]

df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].astype(str)
df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].map(lambda x: x.lstrip('['))
df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].map(lambda x: x.rstrip(']'))
df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].map(lambda x: x.replace('false','"false"'))
df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].map(lambda x: x.replace('true','"true"'))

df_holiday_not_null['holiday_info']=df_holiday_not_null['holiday_info'].map(eval)
holiday_df = pd.json_normalize(df_holiday_not_null['holiday_info'])

processed_holiday=pd.concat([df_holiday_not_null,holiday_df],axis=1)
processed_holiday=processed_holiday.drop(['holiday_info'],axis=1)


processed_holiday




merge_store=pd.merge(df_base,processed_store,on=['date','store_nbr'],how='inner')

merge_holiday=pd.merge(merge_store,processed_holiday,on=['date','store_nbr'],how='left').rename(columns={'type_x':'store_type','type_y':'holiday_type'})
merge_holiday


interim_df=merge_holiday.copy()

interim_df['date']=pd.to_datetime(interim_df['date'], format="%Y-%m-%d")

interim_df['year']=interim_df['date'].dt.strftime('%Y').astype(int)

interim_df['month']=interim_df['date'].dt.strftime('%m').astype(int)

interim_df['day_of_mnth']=interim_df['date'].dt.strftime('%d').astype(int)

interim_df['day_of_week']=interim_df['date'].dt.strftime('%a')

interim_df['daynum_of_week']=interim_df['date'].dt.strftime('%w').astype(int)

interim_df['mnth_name']=interim_df['date'].dt.strftime('%b')



prepped_data.to_csv('store_sales_data_processed.csv',index=False)

prepped_data
