==> 02-1-Basic Findings.ipynb <==
import os
os.chdir('/content/drive/MyDrive/Colab Notebooks/Technical Assessement THD')
print(os.getcwd())
import pandas as pd
import numpy as np
import datetime

import plotly.express as px

df=pd.read_csv('modeled_data.csv')

df['date']=pd.to_datetime(df['date'])
df['week']=df['date'].dt.strftime('%U').astype(int)

df=df.sort_values(['store_nbr','date'],ascending=[True,True])

df.dtypes

df.loc[(df['holiday_type'].isnull()),'holiday_type']='normal day'
df.loc[(df['holiday_type']=='Holiday') & (df['transferred']==True),'holiday_type']='working holiday'

df['promo_band']='none'
df.loc[(df['promotion_count'] >0) &(df['promotion_count'] <=34),'promo_band']='low'
df.loc[(df['promotion_count'] >34) &(df['promotion_count'] <=165),'promo_band']='med'
df.loc[(df['promotion_count'] >165),'promo_band']='high'


df.loc[(df['holiday_type'].isnull()),'holiday_type']='normal day'
df.loc[(df['holiday_type']=='Holiday') & (df['transferred']==True),'holiday_type']='working holiday'


sales_hol=df[df['active']=='Y'].groupby(['holiday_type'])['sales'].mean().reset_index()


fig = px.bar(sales_hol, x='holiday_type', y='sales',text_auto='0.4s')
fig.show()
print('\n')
sales_hol.to_csv('holtbl.csv',index=False)

df.promotion_count.describe()

df['promo_band']='none'
df.loc[(df['promotion_count'] >0) &(df['promotion_count'] <=34),'promo_band']='low'
df.loc[(df['promotion_count'] >34) &(df['promotion_count'] <=165),'promo_band']='med'
df.loc[(df['promotion_count'] >165),'promo_band']='high'

sales_promo=df[df['active']=='Y'].groupby(['promo_band'])['sales'].mean().reset_index()
sales_promo.to_csv('promotbl.csv',index=False)

cluster_cnt=df.groupby(['cluster'])['store_nbr'].nunique().reset_index()
cluster_cnt.to_csv('clustercnt.csv',index=False)
