==> 05-Plot generator.ipynb <==
!pip install -U kaleido

import os
os.chdir('/content/drive/MyDrive/Colab Notebooks/Technical Assessement THD')
print(os.getcwd())
import pandas as pd
import numpy as np
import datetime
import kaleido
import plotly.express as px
import plotly.io as pio
import plotly.graph_objects as go
pio.kaleido.scope.default_format = "png"




##for single store clusters
df=pd.read_csv('single_store_output.csv')

df['date']=pd.to_datetime(df['date'])

df=df.sort_values(['store_nbr','date'],ascending=[True,True])

store_lst=list(set(df['store_nbr']))

for i in store_lst:

  sdf=df[df['store_nbr']==i]

  fig1 = go.Figure()

  fig1.add_trace(go.Scatter(x=sdf['date'], y=sdf['sales'],
                    mode='lines',
                    name='sales'))
  
  t='store'+str(i)
  fig1.update_layout(
    title=t)
  
  
  s='raw_store_'+str(i)+'.png'

  s1='output plots/'

  s2=s1+s

  fig1.write_image(s2)

  #os.chdir('/content/drive/MyDrive/Colab Notebooks/Technical Assessement THD/output plots')
  
  sdf2=sdf[sdf['outliers']=='Y']

  fig1.add_trace(go.Scatter(x=sdf2['date'], y=sdf2['sales'],
                    mode='markers',
                    name='outliers',marker={'color':'red'}))
  
  s='outlier_store_'+str(i)+'.png'

  s1='output plots/'

  s2=s1+s

  fig1.write_image(s2)

  sdf3=sdf[sdf['outlier_type']!='Not outlier']

  fig2 = go.Figure()

  fig2.add_trace(go.Scatter(x=sdf['date'], y=sdf['sales'],
                    mode='lines',
                    name='sales'))
  
  t='store'+str(i)
  fig2.update_layout(
    title=t)



  fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='outliers']['date'], y=sdf3[sdf3['outlier_type']=='outliers']['sales'],
                    mode='markers',
                    name='outliers',marker={'color':'red'}))


  fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='weak outlier with holiday']['date'], y=sdf3[sdf3['outlier_type']=='weak outlier with holiday']['sales'],
                    mode='markers',
                    name='weak with holidays',marker={'color':'cyan'}))
  
  s='final_store_'+str(i)+'.png'

  s1='output plots/'

  s2=s1+s

  fig2.write_image(s2)

  








  


##for multi store clusters

df=pd.read_csv('multi_store_output.csv')

df['date']=pd.to_datetime(df['date'])

df=df.sort_values(['store_nbr','date'],ascending=[True,True])


cluster_lst=list(set(df['cluster']))

for k in cluster_lst:
  df1=df[df['cluster']==k]
  store_lst=list(set(df1['store_nbr']))

  for i in store_lst:

    sdf=df1[df1['store_nbr']==i]

    fig1 = go.Figure()

    fig1.add_trace(go.Scatter(x=sdf['date'], y=sdf['sales'],
                      mode='lines',
                      name='sales'))
    
    t='store'+str(i)
    fig1.update_layout(
      title=t)
    
    
    s='raw_store_'+str(i)+'.png'

    s1='output plots/'

    s2=s1+s

    fig1.write_image(s2)

    #os.chdir('/content/drive/MyDrive/Colab Notebooks/Technical Assessement THD/output plots')
    
    sdf2=sdf[sdf['outliers']=='Y']

    fig1.add_trace(go.Scatter(x=sdf2['date'], y=sdf2['sales'],
                      mode='markers',
                      name='outliers',marker={'color':'red'}))
    
    s='outlier_store_'+str(i)+'.png'

    s1='output plots/'

    s2=s1+s

    fig1.write_image(s2)

    sdf3=sdf[sdf['outlier_type']!='Not outlier']

    fig2 = go.Figure()

    fig2.add_trace(go.Scatter(x=sdf['date'], y=sdf['sales'],
                      mode='lines',
                      name='sales'))
    
    t='store'+str(i)
    fig2.update_layout(
      title=t)



    fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='Strong - no holiday and deviates from cluster']['date'], y=sdf3[sdf3['outlier_type']=='Strong - no holiday and deviates from cluster']['sales'],
    mode='markers',
    name='strong',marker={'color':'red'}))


    fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='Med strong - holiday but deviates from cluster']['date'], y=sdf3[sdf3['outlier_type']=='Med strong - holiday but deviates from cluster']['sales'],
    mode='markers',
    name='med strong',marker={'color':'pink'}))

    fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='Med Weak - no holiday but low deviation from cluster']['date'], y=sdf3[sdf3['outlier_type']=='Med Weak - no holiday but low deviation from cluster']['sales'],
    mode='markers',
    name='med weak',marker={'color':'yellow'}))

    fig2.add_trace(go.Scatter(x=sdf3[sdf3['outlier_type']=='Weak- holiday and minor deviation from cluster']['date'], y=sdf3[sdf3['outlier_type']=='Weak- holiday and minor deviation from cluster']['sales'],
    mode='markers',
    name='weak',marker={'color':'cyan'}))


    s='final_store_'+str(i)+'.png'

    s1='output plots/'

    s2=s1+s

    fig2.write_image(s2)
